pyproject.toml
[tool.poetry]
name = "yin-yang"
version = "3.3.0"
description = "Auto Nightmode for KDE, Gnome, Budgie, VSCode, Atom and more."
authors = ["Oskar Schachtschneider"]
repository = "https://github.com/oskarsh/Yin-Yang"

license = "MIT"
readme = "README.md"
packages = [
  { include = "yin_yang" }
]

[tool.poetry.dependencies]
python = "^3.11,<3.13"
psutil = "5.9.8"
PySide6-Essentials = "6.6.3.1"
shiboken6 = "6.6.3.1"
PySide6-Addons = "6.6.3.1"
suntime = "^1.3.2"
systemd-python = "235"
requests = "2.31.0"


[tool.poetry.group.DEV.dependencies]
flake8 = "^7.0.0"
pytest = "^8.1.1"
pyyaml = "^6.0.1"
toml = "^0.10.2"
setuptools = "^69.5.1"
wheel = "^0.43.0"
cython = "<3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yin-yang = "yin_yang:__main__"

[tool.pyright]
include = ["yin_yang"]
exclude = ["**/node_modules",
  "**/__pycache__",
  "build",
  "pytest_cache",
  ".flatpak-builder"
]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"